requirements.txt
# Requirements for test scripts
requests>=2.31.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-otlp-proto-http>=1.21.0

